+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### [0.3.0] Unreleased
+- Changed: HasRawWindowHandle and HasRawDisplayHandle are now trait bounds on VkInit functions.
+- Changed: env_logger only as dev-dep.
+- Changed: PhysicalDevice now reports all property limits.
+
 ### [0.2.0] Restructure
 - Changed: All shader compilation functionality is now behind the `shader` feature to make shaderc an optional dependency.
 - Changed: Moved examples to separate crate: `vku-examples`.